Move Request queries into new class
